Email User Variable fix for Users module condition
The problem arises when you set a condition based on the Users module in a Workflow. When you try to send an email as an action, the Contact and User variables are not parsed correctly. The reason is the User record is passed as a bean to the run_action() function in the actionSendEmail.php and code on line 244 tries to store the $bean->assigned_user_id to an array that is then passed to the static aowTemplateParser::parse_template() function. The issue is the user record does not have an assigned_user_id property so it is set as NULL and none of the email variable are parsed correctly.