minimum price support
Adds support for minimum price enforcement. When value entered on quote is less than the specified minimum price then user is presented with a dialogue box explaining that the price is too low. This can prevent a salesperson from accidentally quoting a discounted price lower than the business's cost for the product (or below a minimum gross margin).