ProspectList: Related field fix
If you add a custom field to Accounts which it is of type: relate then the Export query for ProspectList is broken.
If insists on using name field as its table value while it should use the ext3 field instead.

TODO: Check if other type of fields need similar workarounds or not.