Fixes #517 The problem is caused by line 842 the strip_tags function.
More specifically when getModuleField returns the field in most cases it returns a value enclosed in span tags so when strip_tags is run it strips off the span tags and leaves the value but in the case of a check box field getModuleField just returns the input html tag with its value checked so therefore strip_tags removes everything,  thus leaving it blank.

There does not seem to be an easy way to fix this without altering aow_utils.php since AOR_Fields doesn’t store the field type. So I have added code to check if it is a checkbox directly in the returned string and whether its been checked or not.