FR 3.0, 3.6 , 3.7, 3.8 , 3.12
added contacts and user relationship changes, added gant chart view.

As a User I want the CRM to display Project Template information in the
Project Template Gantt view in the generic SuiteCRM format so that I do
not get confused by the layout.

As a User I want to click on a button which says 'Add A New Task'which
will add a new Project Template Task to the Gantt Chart.

As a User when I am in the Gantt View of a Project Template I want to
click on the View Details button which will be displayed at the top
right of my screen so that I can easily navigate to the detail view of
my Project Template.

As a User I want to see the Resource column in the Project Template
Gantt view display resources by their (full name, preferred name??) so
that I can fully understand who is a resource for the project.