Strip HTML tags from notification e-mail Subjects
Email subjects should never include HTML.
Since there is a difficulty with translated email templates that is causing HTML tags in email "Subjects" to be visible to end-users, and we couldn't find any way to fix it in Crowdin, this change solves the problem.
At the same time, it's follows a good general principle - HTML tags don't belong there.