US 2.1, US 2.4, US2.7
US 2.1 - As a User I want Gantt views to display the appropriate drop
down
action buttons so that I can access functionality on the Gantt View.

US 2.4 - As a User I want to be able inline edit fields on the Gantt
views so that it
follows generic SuiteCRM functionality.

US 2.7 - As a User I want the detail view section of the Gantt view to
be
consistent with the rest of the CRM so that I can in line edit and
access
action buttons.