Optimistic Locking not working in some cases
In some cases sugar retrieves more object of the edited type. This leads to that the optimistic locking id is not the one thats in the edit view, but some other (maybe from relation).
So if someone saves the same record in a different session, the one that saves later simply overwrites the first changes, and does not show the dialog for conflicting saves.
This change makes sure the id from the request is stored in the session, not the id from the last fetched record that maybe from a related bean.