Project Duplicate Problem fixing
When duplicating project, the duplicate is happening but the project tasks related to that particular project is not copying . instead of that some other tasks are copying to duplicated project.  I found that duplicateid is parsed to Save.php but Save.ph is looking for relate_id which is a bug.