FIXED - PHP Notice: A non well formed numeric value encountered #3577
PHP Notice: A non well formed numeric value encountered in SUITECRM/include/ListView/ListViewSmarty.php on line [102|108] #3577

The problem here a different themes/templates using PIXEL or PERCENT as width.
This fix checks which one is used and decides what to do. 
In case PIXELS are used some adjustments are required
in case PERCENT is used nothing needs to be adjusted.