Fix connection to dovecot imaps server
SuiteCRM was failling to connect to the server, because findOptimumSettings
was erroneously assuming the credentials were incorrect and stopping the
loop early.

The test failed on the service=imap/ssl/validate-cert/secure test, with error
'[CLOSED] IMAP connection broken (server response)', but succeeded on the
service=imap/ssl/validate-cert test.

The dovecot server was running with:
- disable_plaintext_auth = yes
- auth_mechanisms = plain login
- ssl = required
- ssl_protocols = !SSLv2 !SSLv3