New approach - add SuiteP if needed
I don't know if this approach is better than the one in the previous commit, but at least it doesn't break the app, and doesn't force an admin to take immediate action. 
I didn't try the effect of this on the upgrade process yet - I'm not sure if this allows the upgrade to complete successfully - that would be really nice.