Workflow Enhancements 7.9.1 - Process Once Per Day
Conditions when this is an issue:

Time-based workflow condition which triggers on a particular day. On that day the conditions are met all day and the workflow will trigger every time the schedule runs (e.g. every minute). Repeated runs is required as the date may change to another value in the future and the record should be considered to be processed in the workflow again.

Proposed Solution:

Add checkbox next to the Repeated Runs checkbox called 'Trigger Only Once Per Day'.
When this checkbox is ticked, once it has run once for a particular record it will not run again for 24 hours for that particular record. There is also an Hour & Minute field to specify at what time each day this workflow should run.

REF: https://github.com/salesagility/SuiteCRM/pull/3803