Fixed #4213 - Cannot edit Layouts in Module Builder
The getFileName function in UndeployedMetaDataImplementation is
different from the other implementations, since it allows using
alternate packages. Since it is no longer a static function, separate it
into getFileNameInPackage and change getFileName to assume the
instance's package, so it is compatible with the base class.

This should bring it into line with the usage of getFileName in the
other implementations.

Also, repair the broken phpdoc on getFileName.