Variables missing from email template sent from worklow
When sending an email from a workflow, the template contains variables that are decimals. When the email is received these variables are not populated.

When the email template replace variable functionality, there is no option to replace a decimal field type. There is currency, dropdown, multi dropdown, int, image available but decimal type does not. I have added decimal field type. 

We have fixed that.