Fix #533
Updated date assignments to not use fetched_row as when the date/datetime values are pulled from a newly created record the values do not yet exist in the database and so fetched_row returns false. This causes the date/datetime value to be blank which results in the value being set to today/now. New assignment gets the values straight from the bean rather than the database which ensures the values are set correctly.