Fix #206
Added conditional statement to check if the table_name is campaign_log and if so refactor queries slightly to use activity_date instead of date_entered on both new and modified records to prevent query failure. Previously the query failed when workflows were setup to be triggered by Campaign Logs as the date_entered field does not exist in that table and instead there is an activity_date field.