Fix LDAP login for LDAP servers which require a full DN for binding
ldap_rdn_lookup() tries to bind with the username, the admin username and finally
anonymously, and then gives up. If that succeeds and it finds the username it will
try to bind with the username and if that fails with the full DN.

In our case (Microsoft LDAP) the server only allows binding with a full DN and not
anonymously, so ldap_rdn_lookup() fails and the full DN is never tried.

To fix this try binding with the full DN everywhere instead of just at the end.