Email Optimisation: only processes fields in viewdefs
Issue:
SugarCleaner::cleanHtml is being called twice, once for the html
version of the body, once for the plain version of the body.

Solution:
Since we do not need to display both versions of the body, we car
process the body we need and reducing the cleanHtml to be called only
once.

(cherry picked from commit fc1161f)