templateParser.php - render 'wysiwyg' fields in PDFs
This makes fields from thw 'wysiwyg' addon field render in PDFs
properly.
See issue #241 which is currently closed.
This commit is the 2nd one I've done. The first was to my
'master' branch and had to be reverted and put into it's
own branch, in the hopes that it will be accepted upstream.