Fix quicksearch for member org in accounts
Quicksearch for the member organization (parent_name) field in Accounts does not work.  The quicksearch dropdown list works but does not attempt to populate the related _display field.  The conditional code for parent_name does not work correctly for this field.  It works correctly without it.  (This might mean the better fix would be to change how the field is built so it works like the other parent_name fields...)

Found in 7.8.19 - verfied exists in demo 7.10.6.

#### Expected Behavior
<!--- Tell us what should happen -->
Quicksearch should correctly populate the member organization field in Account Detail View.

#### Actual Behavior
When selected the field will goes blank.

#### Possible Fix
See code change.

#### Context
Grumpy Sales Exec showed me the bug while attempting to enter something.

#### Your Environment
<!--- Include as many relevant details about the environment you experienced the bug in -->
* SuiteCRM Version used: 7.8.19
* Browser name and version (e.g. Chrome Version 51.0.2704.63 (64-bit)): Firefox 52.3
* Environment name and version (e.g. MySQL, PHP 7): PHP 7.1
* Operating System and version (e.g Ubuntu 16.04): CentOS