allow users to email (#6)
* Add option: allow users to send email as themselves w/ System outgoing

With this option enabled, users will be able to send email out
of SuiteCRM with their name and email address, but sending though
the system outgoing email server. Their email address and name
will be populated in the "From" dropdown.
Clean up the text, and shorten the LBL_ALLOW_DEFAULT_SELECTION
label so it doesn't mess with the formatting.