Fixed #6487 - opt-in: use the DB time for writing confirm_opt_in_*date
confirm_opt_in_sent_date, confirm_opt_in_date and confirm_opt_in_fail_date get set to
the local time and then saved to the database. Other modules like reports and workflows
expect dates in the DB to be in the DB timezone and convert them to the local time.
Since the dates are saved in the local time this result in the wrong time being used/shown.

Fix this by always using the DB time when setting those fields like everywhere else.