Fixed #4123 - reports: fix adding parameterized reports to target lists
Given a report with parameters, the action "Add To Target List" did not
take into account the report parameters, resulting in all records getting
added to the target list as if there were no parameters.

Fix this by updating the form with the current parameters and also send
it to the "addToProspectList" controller action. In the controller extract
the parameters before creating the query like with other actions.

This makes the list of records added to the target list match the one the user
sees on the report page.