Workflow: Properly delete records which are marked as deleted
There is currently no action for deleting a record, only for modifying one.

The only way atm is to set the "deleted" field during the workflow action which
seems to work but leaves tangling relations in the DB and not all code handles that
properly, leading to weird errors/states and a bloated DB

Improves the current way by checking if the delete field was set during the action
and properly delete the record after the action is done, including relations etc.