Fixed #6504 - Fix and implement proper bounce email handling
Bounce handling is currently broken for multiple reasons:

(1) The imported bounce email is saved in the notes description field which is
marked as text and strips away all HTML on save(). This removes/mangles the
link containing the tracking ID so the attachment can't be linked to the
campaign log later on.

(2) SMTP servers usually return the original email as an attachment on a
bounce and include an addition attachment containing the delivery status (RFC
3464). Since this attachment wasn't saved, hard bounces weren't detected.

(3) In case the tracking ID makes it through and the email is considered a
hard bounce then marking the email as invalid failed because
markEmailAddressInvalid() was called with an email object but expected a
string.

To fix all these:

For (1) we don't save the attachment in the description and read it from the
saved attachment file instead. This makes sure we get the original content and
can extract the tracking ID.

For (2) we also save the delivery-status attachments in addition to the email
copy and also evaluate it in the bounce handling code. This allows us to
confidently classify hard bounces by parsing the status and diagnostic fields
in the delivery status attachment.

For (3) pass an email address to markEmailAddressInvalid()

Fixes #6504