Fixed 6472 - Fix wrong image sizes for email templates in the campaign wizard
In the campaign wizard the editor (mozaik in this case) is embedded in the main
table. The table has the CSS classes "other view" and the theme contains various rules
that just force the size of all img tags there etc.
This affects various elements of the editor making the preview not match the actual email.

To avoid the theme CSS affecting the editor move the editor out of the table.

Fixes #6472