Add a SAML2 metadata endpoint (named "SAML2Metadata")
This adds a new endpoint which returns the XML metadata for the saml integration.
The output can be used to register the service with the identity provider.

The endpoint can be reached through "index.php?entryPoint=SAML2Metadata"

Based on the php-saml example: https://github.com/onelogin/php-saml/blob/master/demo1/metadata.php