campaing wizard: fix template editor being shown in all steps
In 442af7f8539d99c2 the template editor was moved out of the main table to prevent
the table css affect the template editor content.

I missed that the content in the table is replaced in the wizard steps and by moving
the template editor out the editor is now being shown in all steps instead of only the
first one.

Fix this by tying the visiblity state of the editor to the first step panel.