Fixed #3598 - WebToPersonCapture: Fix assignment of multi select values
In case the the web form contains a multi select field the resulting
value is passed as an array. This was assigned as is to the newly created
record and probably thrown out at some later point because it wasn't a string.

To fix this use encodeMultienumValue() to convert the value list to the expected
internal format and assign that instead.