Also set opt-in send/fail dates when sending emails through the action menu
In case one uses "Send Confirm Opt In Email" in the "Actions" menu of a contact/lead etc
then action_sendConfirmOptInEmail() gets called which in turn calls EmailMan::sendOptInEmail().
This leads to the email being sent without any send/fail dates being recorded.

To fix this set the dates in EmailMan::sendOptInEmail() instead of leaving it to the caller.

Also remove the date set code from WebToPersonCapture because that's no longer needed now that
sendOptInEmail() handles it.