Fixed #6484 - Fix image insertion with mozaik
In #6329 mozaik was set to a dummy object to make it work when not using
the mozaik editor. But the added variable declaration shadows the global
variable breaking the case where mozaik is actually used, setting it to
undefined.

This breaks inserting images into emails, the file gets uploaded but the
URL field stays empty.

Fix this by explicitly checking if we are using mozaik instead of using
a dummy object.