EmailTemplates: Use html2text library for generating the text only body
When creating an email template with the editor and saving then SuiteCRM
will generate a text version automatically by stripping away tags and decoding
html entities (it also gets regenerated if the text field is cleared before saving)

This currently has the downside that many features of the email are lost,
like basic structure through headings/lists and, more importantly, links.

This patch adds the soundasleep/html2text library and uses it instead to generate the
initial text version. It also adjusts the regex for separating variables from surrounding
text to be less strict to allow using variables in HTML links (see the tests)

This for example makes the generated text from the included opt-in confirmation email usable
without having to manually fix it.

Also adds a few tests for common edge cases.