Reports: fix "Illegal string offset 'record_id'" errors.
The 'params' value was changed from an array to a string in 6bdf937e9a07cd55e8
but not all usages fixed. This results in the "$att['params']['record_id']" assignment failing
and in lots of "Illegal string offset 'record_id'" errors when viewing a simple report.

This patch renames the 'params' key to 'format', because it's used only for the field format string
at this point, and creates a new $params array for passing the record_id to getModuleField().
From what I see the field format is checked everywhere because the reports module implements
its own date formatting and avoids calling into getModuleField() when a field has its own
date format set.

This gets rid of the "Illegal string offset" errors and results in the correct record_id getting
passed to getModuleField().