Reports: Fix "One of" operator for multi select fields
MultiSelect fields are encoded like "^foo^,^bar^" in the DB so we need
to encode the values  we want to match against first and have to use
LIKE to match a substring because the field can contain multiple values.

This just copies the logic from the WorkFlow module where this was already
implemented this way:

https://github.com/salesagility/SuiteCRM/blob/ad0f598cd6d561c92f2f8d3ebeae519a12472e6f/modules/AOW_WorkFlow/AOW_WorkFlow.php#L523

This does not fix the other operators which are still broken.