Reports: Hide inaccessible modules in the reports editor
The reports module already hides all fields (in the field list in the bottom left corner)
for modules not accessible to the user but still shows them in the "Module tree" and the
"Report Module" dropdown.

This leads to the users seeing lots of modules and potential relations which are not
usable in the end. To make this clearer also hide inaccessible modules in the module tree
and the module dropdown.

The call to load_report_beans() is moved to fill_in_additional_detail_fields() because
calling into ACLController() in the constructor results in an infinite recursion as it
instantiates a AOR_Report() itself.