Fixed #6869 - Users: fix saving user records resetting user preferences
In #6404 the code to apply the user edit view form to the preferences
was moved into User::save() unconditionally. This results in the preferences
being reset in case save() is called in a different context than the form.

For example when inline editing a user in the user list view the preferences
for that user get set as if there was an empty edit form.

Fix this by only applying the form updates in case we are in the context of
saving in the edit view page.