Fix Content-Type header missing in some cases for the getImage entry point
The getImage entry point only handled setting the correct mime type for
gif and png images. In case of svg images (for example the icon left of attachments in
the email template editor) the returned content type was text which made firefox
show an error and not display the image.

Fix this by adding mime types for all supported image extensions in SugarTheme
and using that in the getImage end point.

This also adds some basic tests for this new API.