phpunit: Reduce test output
phpunit provides helpers for testing and redirecting test output but this is broken
by the added code in the test base class which prints test context during the tests
and gets treated as test output by phpunit.

To work around this write straight to STDOUT instead of going through the output buffer
for the test context.

This also adds some calls to setOutputCallback() and expectOutputRegex() to remove
the output/noise of some tests.