Links returned in the API should be relative to the base_uri
fixes #7095
The new API states the base_uri should `https://crm.example.org/Api/`.
This means all links returned from the API need to be relative to this
`base_uri`.

According to [Guzzle explanation of RFC
3986](http://docs.guzzlephp.org/en/stable/quickstart.html#making-a-request)
the `base_uri` with a path should end with `/` and the relative link
should start without a `/`.

Full Example usage of this would

```php
<?php
$client = new \GuzzleHttp\Client([
    'base_uri' => 'https://crm.example.org/Api/',
]);

$url = 'V8/module/Notes?page[size]=20';
do {
    $response = $client->request('GET', $url);

    $body = json_decode((string)$response->getBody());

    $data = $body->data;
    // Do something with data

    $url = property_exists($body, 'links')
            && property_exists($body->links, 'next')
            && !is_null($body->links->next) ? $body->links->next : null;
} while ($url);
```