Add php-cs-fixer to composer.json as a dev dependency.
The composer.json didn't specify php-cs-fixer despite the fact that we
use it for the Robo style check command. This commit adds it explicitly
to the composer.json so it will be installed by default for developers
and so it can be used with the same specific version for everyone.