Add CKEditor 4 as a new email template editor option
This adds ckeditor through composer and integrates it as an email
template editor. It can be activated in the user profile under
"Email Setting" > "Editor".

This enables various plugins which are useful for composing fully
formatted text, tables, working with images etc.

This also copies the template functionality from the existing "Mozaik" editor.
The template button shows a preview of the provided templates. They are
very basic and based on tables though (but tested with Outlook).

And finally this also adds support for image upload, like with Mozaik.