Image not found because "s" is added
I do not know, if anybody else has this.
For me, Suite tries to load an  image "Meetingss.gif". So double ss in the end. The gif of course is named "Meetings.gif"
I know that at some places (at least in the past), plural of module names were handled differently, but do not know if that is the reason for the additional "s".