travis-ci: Switch the PHP 7.x jobs to Ubuntu 16.04
Currently all tests are running with Ubuntu 14.04 which includes MySQL 5.6
which is unlikely to be used by any newer setup.

Assuming that systems with a newer PHP version also use a newer MySQL switch the
PHP 7.x jobs to Ubuntu Xenial and thus to use MySQL 5.7 instead.

The unittests currently have problems with the default MySQL 5.7 configuration,
see #7107, so force unset the strict sql mode in the travis setup script for now.

This should make it easier to test the fix for #7107 and to prevent similar issues
in the future.