Fixed #6420 - Campaigns: don't send test emails to lists no longer linked to the campaign
When querying the emailman table the code didn't take deleted records and links into
account which could result in test emails being sent to persons no longer linked to
the campaign.

This adds deleted checks to the query.
Also adds a deleted check for the emailman entries for both test and non-test cases
while at it.

Fixes #6420