Fixed #7107 - MysqliManager: always remove STRICT_TRANS_TABLES from sql_mode
STRICT_TRANS_TABLES is enabled by default with MySQL 5.7 and some tests
fail with it due to missing values in SQL insert/update statements.

Given that our test coverage isn't that high there might be more issues
not covered by the tests, so better disable STRICT_TRANS_TABLES to get the same
behaviour as with MySQL 5.6 for now.