Remove implicit waiting in acceptance tests on Travis CI.
This adds a bunch of waitForElementVisible, waitForText, etc. where
necessary and will theoretically make the test suite more consistent
between machines of different speeds. The wait time in the travis
codeception env causes problems if you run the test suite on a slower
machine because it causes tests that assume they have x number of
seconds to wait for an expectation (like `I->see`, or `I->click`), when
that amount of time can vary depending on the machine.

Instead of increasing the wait value for slower machines, I'd prefer to
keep it consistent across machines so we can run CI anywhere without
needing to configure as much. This commit moves the waiting from the
codeception env to being in the tests explicitly.

Could also speed up the test suite.