CI-REFACTOR: Passwords are madatory
The mariadb container's `MYSQL_ROOT_PASSWORD` variable is required.
Larter this should be updated to either let the container generate a
random one, or have the option for the build pipe pass one.

We're also adding support for a mysql user password, currently
one can either be passed or not, allow empty passwords is on.