Scheduled Reports: Enable security groups support and add the subpanel
Without security groups when a user creates a new scheduled report then no
other user in the same security group can see/edit the record.

This adds security group support to the module and also adds a security groups
subpanel in the detail view so the group relations can be edited.