ListView: Fix selection when switch from "select all" to "select page"
When selecting all list entries and switching between pages and then switching the selection
back to "select page" the code tried to keep the current selection, but this doesn't really work
because the saved uids don't include the whole list, so this results in a random selection depending
on the pages visited between the two selection actions.

1) Select all
2) Next page -> previous page
3) Select page
4) Now the first two pages are selected, which is neither "all" or "page"..

Instead keep it simple and when going from "all" to "page": clear the selection state and select the
current page.