export: Update export_excel_compatible to work with all Excel versions
Excel on Windows/macOS when simply opening a CSV file will try to guess the
CSV dialect (encoding, delimiter, etc) and just go with the first guess.
It's still possible to do an explicit import with Excel where you can select the
dialect, but that's not very obvious.

The most common wrong guess is that mostly ASCII utf-8 will get detected as latin-1.

I've tested various encodings with Excel (2019) and codepoints from latin-1:

* utf-8: win/macOS use latin-1
* utf-8+bom: win/macOS correctly use utf-8
* utf-16le: win uses utf-16le, macOS garbled text
* utf-16le+bom: win/macOS fail to detect CSV and load it as one text block

thus utf-8+bom is the preferred mode.

This patch makes the "export_excel_compatible" mode use utf-8+bom for the CSV encoding.

In addition this removes the hardcoded delimiter in export_excel_compatible mode, because excel
uses the Windows user list separator for this and there is no way of knowing it (except in javascript maybe)
and so we better leave this setting up to the user.