EmailMarketing: Add security groups support
When a user of a security group creates a marketing entry then no
other user in the same group can see/edit it.

This adds the security group relation so that new records are associated with
the right group and other users in the same group can access the marketing email
record.