Robo: Add a --filter option to tests:unit for filtering tests
This allows one to run only one specific test method of a test class, which is
helpful when you implement a new test or want to investigate a specific failing one.

The --filter option is passed as is to phpunit, see https://phpunit.readthedocs.io/en/latest/textui.html
for details.