Remove SugarXHProf.
The XHProf project has been abandoned for 4 years, and
the version vendored here is much older than even that. It's
undocumented, unmaintained, and it doesn't seem like it even works
on PHP 7.x.

"The best way to choose what to keep and what to throw away
is to take each item in one’s hand and ask: 'Does this spark joy?'
If it does, keep it. If not, dispose of it."

- Marie Kondō