Add a test for the httpTime() function
I tried to test it with the current time by manipulating 'now',
but the httpTime function just uses PHP's time() function, and
changing it to call  getNow() is annoyingly difficult.

I wonder if we should use a Composer package for this purpose. Ruby has a gem called Timecop that does this kind of thing.