FIX ControllerFactory
ControllerFactory crashes when the custom controller
class name is wrong. In other cases, we already default to
the SugarController. This is now the default for all
cases where a custom controller could not be loaded
for any reason. Whenever a custom controller file is
found, but the controller can't be instantiated,
appropriate log messages are emitted.