Fix the user factory causing a warning.
Something about needing both a username and email, for some reason
it wants the user to have a hashed password set? I'm not sure
if that warning message being printed in that case is a bug or what.

Should probably figure out a way to allow the user to pass the password
into the factory, that way it can be used in tests easily.

For now, it just uses a random one from Faker, which can be overridden
if necessary.