Fix invalid SQL generated from SOAP call to get_entry_list for Emails
$where_auto was empty when $show_deleted was neither 0 or 1
(e.g.  service/v4/SugarWebServiceImplv4.php sets it to -1 to signify
"both deleted and non-deleted objects").

However the code in Emails.php would generate an invalid SQL query
if $where_auto was empty:
        if ($where != "") {
            $query .= "WHERE $where AND " . $where_auto;
        } else {
            $query .= "WHERE " . $where_auto;
        }
.. which leads to a query ending in "AND" with nothing after that.
Obviously this leads to "You have an error in your SQL syntax [...]"

The solution of initializing $where_auto to "1=1" is the one used
in other files like Opportunity.php