theme SuiteP: Use LatoLatin instead of Lato
Some characters used in Czech like ěčř or Romanian language are not present in
Lato font and they are rendered in Arial, which does not look good.

There exists LatoLatin [1] from the same author, which includes glyphs
for European Latin-based languages.

More details can be seen in my pull request on GitHub.

[1] https://www.latofonts.com/lato-free-fonts/