Fix #9203 - Duplicate notification emails.
Prevent duplicate notification emails being sent.
When changing the assigned to user and the module has a workflow
with a Calculate Fields action with a formula and the workflow Repeated Runs option is selected.
Add additional flag to set to true when the notification email has been sent.
Add additional check before sending the notification email.