Make error response comply with JSON:API (OpenAPI 3.0)
Presently we do not comply with JSON:API specs, see:
  https://jsonapi.org/format/#error-objects

This leads to incompatibility with proper PSR-7 implementations
such as https://github.com/Nyholm/psr7 (recommended by Symfony).

  "Warning: array_key_exists() expects parameter 2 to be array, int given"

Sadly we cannnot use excellent JSON API Client packages such as:
  https://github.com/eosnewmedia/JSON-API-Client

It is not possible to extend the deserializer because at this point
we lost the required array key (e.g. 'status').

Thus do it the proper way and make our error response compatible with specs.

TODO (maybe): There is another issue with the prefixed Api V8 and 'module'
string but this issue can be worked around.