Conditions do not work on_save
When saving a record AOW does call check_valid_bean in the logic hook
In line 828 following line does not compare to "false" and as of that if compare_condition returns "0" it interprets it as "false"
 if (!($this->compare_condition($field, $value, $condition->operator))) {

In compare_conditions we give back the result of strpos. This is "0" if the comparision is true, but for the first char.
Example:
Fieldvalue: Myspecialfieldname
Condition: Starts_With "M"
Should be true, Record should run the Workflow, but does not run because strpos gives back "0" interpreted as false.