Update Delete.php to mark group folders as deleted RE:#9028
When group email accounts are deleted the associated folders don't get marked as deleted=1.  This means they show up for users in folder selection.  They should be marked as deleted when the inbound group email account is deleted.  There is in fact, functionality to do this in SugarFolders.php but is is not called when an inbound group email is deleted.  This additional code just calls the existing function delete() when an inbound email account is deleted.