Fix #5511 -  Calculated fields doesn't handle numeric values with non US locale
When using non US locale and using a mathematical caluclation it returns the string in the wrong number format, which on save will result in incorrect calculations.

Example calc:

`multiply({P0};{P1})`

if P0 is 0.8 and P1 is 0.5 the calculated field will set the string `"0.4" ` to the selected field. If this is a numeric field on save it will parse the result. In german local the dot is the thousand separator, resulting the field beeing saved as 4 and not 0.4 in the database.